+++ clang/lib/Lex/Preprocessor.cpp
@@ -1438,7 +1438,11 @@ void Preprocessor::EnterModuleSuffixTokenStream(ArrayRef<Token> Toks) {
 void Preprocessor::makeModuleVisible(Module *M, SourceLocation Loc,
                                      bool IncludeExports) {
   CurSubmoduleState->VisibleModules.setVisible(
-      M, Loc, IncludeExports, [](Module *) {},
+      M, Loc, IncludeExports,
+      [&](Module *Imported) {
+        for (const auto *File : Imported->Includes)
+          CurSubmoduleState->VisibleIncludedFiles.insert(File);
+      },
       [&](ArrayRef<Module *> Path, Module *Conflict, StringRef Message) {
         // FIXME: Include the path in the diagnostic.
         // FIXME: Include the import location for the conflicting module.